+++ src/mode_predict.py
@@ -46,6 +46,12 @@ if __name__ == "__main__":
         omicLogger.info("Checking for Trained models")
         model_path = assert_best_model_exists(experiment_folder)
 
+        # if the data is R2G then warn the user that the prediction data must have been pre-processed the same way
+        if config_dict["data"]["data_type"] == "R2G":
+            omicLogger.warning(
+                "Previous model was trained with ready to go data. Please ensure that the data being given to this mode has been pre-processed in exactly the same way."
+            )
+
         omicLogger.info("Loading Data...")
         x_to_predict, _, features_names = utils.load.load_data(
             config_dict, mode="prediction"