+++ src/utils/parser/data_model.py
@@ -4,11 +4,7 @@
 # https://opensource.org/licenses/MIT
 
 from typing import Literal, Union
-from pydantic import (
-    BaseModel,
-    FilePath,
-    DirectoryPath,
-)
+from pydantic import BaseModel, FilePath, DirectoryPath, model_validator
 
 
 class DataModel(BaseModel):
@@ -20,5 +16,12 @@ class DataModel(BaseModel):
     save_path: DirectoryPath = "/experiments/"
     target: str
     data_type: Literal[
-        "tabular", "gene_expression", "microbiome", "metabolomic", "other"
+        "tabular", "gene_expression", "microbiome", "metabolomic", "other", "R2G"
     ]
+
+    @model_validator(mode="after")
+    def check(self):
+        if self.data_type == "R2G":
+            self.metadata_file = None
+            self.metadata_file_holdout_data = None
+        return self